Download data from KGS Master List of Oil and Gas Wells in Kansas by choosing township, range, section, etc.

Download data from KGS Master List of Oil and Gas Wells in Kansas by choosing township, range, section, etc.

ogw(township = NULL, range = NULL, range_dir = "W", section = NULL,
  lease = NULL, operator = NULL, state = "kansas",
  county = "any county", api = NULL, welltype = "ALL", all = FALSE)

Arguments

township
Township

range
Range

range_dir
Range Direction

section
Section

lease
Lease Name

operator
Operator Name

state
State

county
County

api
5 digit API Well Number

welltype
Well Type

all
Should we download all data for default values

Value

Tibble data

Examples

# Download data for given township, range, section, etc.

# NOT RUN {
ogw(township = "1", range = "1", range_dir = "W", section = "1")
# }